Change the type of most numeric entities of CPPv2 to 256-bit
The list of changed entities:
* numeric parameters of all public functions;
* all public numeric constants;
* numeric parameters of all events;
* numeric parameters of all errors;
* numeric fields of all internal structs.

It decreases the gas usage and the size of the CPPv2 contract

+++ contracts/interfaces/ICardPaymentCashbackV2.sol
@@ -29,7 +29,7 @@ interface ICardPaymentCashbackV2 is ICardPaymentCashbackV2Types {
      * @param oldRate The value of the old cashback rate.
      * @param newRate The value of the new cashback rate.
      */
-    event SetCashbackRate(uint16 oldRate, uint16 newRate);
+    event SetCashbackRate(uint256 oldRate, uint256 newRate);
 
     /**
      * @dev Emitted when a cashback send request succeeded.
@@ -122,7 +122,7 @@ interface ICardPaymentCashbackV2 is ICardPaymentCashbackV2Types {
      *
      * @param newCashbackRate The value of the new cashback rate.
      */
-    function setCashbackRate(uint16 newCashbackRate) external;
+    function setCashbackRate(uint256 newCashbackRate) external;
 
     /**
      * @dev Enables the cashback operations.
